# nixos/boot/tmp: introduce adaptive huge memory pages [perf]