@paepckehh 2025-05-05
Makefile to verify / benchmark:
```code
#!/bin/gmake
prep:
        ( cd /tmp && git clone https://github.com/nixpkgs && cd nixpkgs && git gc --aggressive )
disk:
        ( touch /tmp/test && rm /tmp/test && dd if=/dev/zero of=/tmp/test oflag=direct bs=16M count=256 && rm /tmp/test )
checkout:
        time ( cd /tmp/nixpkgs && touch test && rm -rf * && git checkout .  && rm -rf * )
hyperdisk:
        hyperfine --warmup 5 "make disk"
hypercheckout:
        hyperfine --warmup 5 "make checkout"
```