@nixos-discourse 2025-05-06
This pull request has been mentioned on **NixOS Discourse**. There might be relevant details there:

https://discourse.nixos.org/t/tmpfs-iommu-h-w-support-via-huge-pages-memory-alloc/63930/1